move commands/utils/administrative_areas.py to services/administrative_areas.py
