+++ app/batid/management/commands/publish_opendata.py
@@ -2,7 +2,7 @@
 from django.core.management.base import BaseCommand
 
 from app.celery import app
-from batid.management.commands.utils.administrative_areas import dpts_list
+from batid.services.administrative_areas import dpts_list
 
 
 class Command(BaseCommand):
