+++ app/batid/services/imports/import_bdtopo.py
@@ -19,7 +19,7 @@ from batid.models import Building
 from batid.models import BuildingImport
 from batid.models import Candidate
 from batid.services.candidate import create_inspection_tasks
-from batid.services.france import dpts_list
+from batid.services.administrative_areas import dpts_list
 from batid.services.imports import building_import_history
 from batid.services.source import BufferToCopy
 from batid.services.source import Source
