+++ app/batid/services/imports/import_dpt.py
@@ -4,8 +4,8 @@ from django.contrib.gis.geos import GEOSGeometry
 from django.contrib.gis.geos import MultiPolygon
 
 from batid.models import Department
-from batid.services.france import fetch_departments_refs
-from batid.services.france import fetch_dpt_cities_geojson
+from batid.services.administrative_areas import fetch_departments_refs
+from batid.services.administrative_areas import fetch_dpt_cities_geojson
 
 
 def import_etalab_dpts() -> None:
